+++ changelog.md
@@ -3,6 +3,8 @@
 
 
 ## Added
+* Perlin noise can now use time as the 3d dim
+* Sample new seed at key press
 * Trees
 * User Control for camera rotation
 * Sea